+++ wheel/wheelfile.py
@@ -10,8 +10,11 @@ from collections import OrderedDict
 from distutils import log as logger
 from zipfile import ZIP_DEFLATED, ZipInfo, ZipFile
 
-from wheel.cli import WheelError
 from wheel.util import urlsafe_b64decode, as_unicode, native, urlsafe_b64encode, as_bytes, StringIO
+
+
+class WheelError(Exception):
+    pass
 
 # Non-greedy matching of an optional build number may be too clever (more
 # invalid wheel filenames will match). Separate regex for .dist-info?
